requirements.txt
tinydb>=4.8.0
openai>=1.0.0
schedule>=1.2.0
pystray>=0.19.0
plyer>=2.1.0
Pillow>=10.0.0
typer>=0.12.0
pytest>=8.0.0
freezegun>=1.5.0
python-dotenv>=1.0.0
platformdirs>=4.0.0
filelock>=3.14.0
